[subtitle] fix blending issue use wrong data type

+++ lib/gdi/accel.cpp
@@ -274,7 +274,7 @@ int gAccel::blit(gUnmanagedSurface *dst, gUnmanagedSurface *src, const eRect &p,
 #ifdef BCM_ACCEL
 	if (!m_bcm_accel_state)
 	{
-		unsigned long pal_addr = 0;
+		unsigned int pal_addr = 0;
 		int src_format = 0;
 		if (src->bpp == 32)
 			src_format = 0;
@@ -286,7 +286,7 @@ int gAccel::blit(gUnmanagedSurface *dst, gUnmanagedSurface *src, const eRect &p,
 			{
 				/* sync pal */
 				pal_addr = src->stride * src->y;
-				unsigned long *pal = (unsigned long*)(((unsigned char*)src->data) + pal_addr);
+				unsigned int *pal = (unsigned int*)(((unsigned char*)src->data) + pal_addr);
 				pal_addr += src->data_phys;
 				for (int i = 0; i < src->clut.colors; ++i)
 					*pal++ = src->clut.data[i].argb() ^ 0xFF000000;
